Created a reusable form component and refactored CreateStream component.#5
