+++ client/src/components/streams/EditStream.jsx
@@ -9,7 +9,12 @@ class EditStream extends Component {
   render() {
     console.log(this.props)
     if (!this.props.stream) return <h4>Loading...</h4>
-    return <h4>Title: {this.props.stream.title}</h4>
+    return (
+      <div>
+        <h1>Edit Stream</h1>
+        <h4>Title: {this.props.stream.title}</h4>
+      </div>
+    )
   }
 }
 const mapStateToProps = (state, ownProps) => {
